+++ src/backend.ts
@@ -113,6 +113,8 @@ export class Backend<M extends IAnyModelType> {
       );
     }
     const processResult = await this.processAll(this.node);
+    this.errorValidations.clear();
+    this.warningValidations.clear();
 
     const completeProcessResult: ProcessResult = {
       updates: [],